+++ src/babel/transformation/transformers/internal/validation.js
@@ -25,3 +25,16 @@ export function Property(node, parent, scope, file) {
 }
 
 export { Property as MethodDefinition };
+
+export function BlockStatement(node) {
+  for (var i = 0; i < node.body.length; i++) {
+    var bodyNode = node.body[i];
+    if (t.isExpressionStatement(bodyNode) && t.isLiteral(bodyNode.expression)) {
+      bodyNode._blockHoist = Infinity;
+    } else {
+      return;
+    }
+  }
+}
+
+export { BlockStatement as Program };