+++ test/fixtures/transformation/es6-classes/preserves-directives/actual.js
@@ -1,0 +1,6 @@
+class MyCtrl {
+  constructor(a) {
+    "any directive prologue";
+    foo;
+  }
+}