+++ test/fixtures/transformation/es6-classes/preserves-directives/expected.js
@@ -1,0 +1,7 @@
+"use strict";
+
+var MyCtrl = function MyCtrl(a) {
+  "any directive prologue";
+  babelHelpers.classCallCheck(this, MyCtrl);
+  foo;
+};