+++ CHANGELOG.md
@@ -17,6 +17,7 @@ See [CHANGELOG - 6to5](CHANGELOG-6to5.md) for the pre-4.0.0 version changelog.
 
  * **Bug Fix**
   * Rewrite named function expressions in optional async function transformers.
+  * Hoist directives.
  * **Internal**
   * `spec.typeofSymbol` transformer has been renamed to `es6.symbols`.